+++ packages/web/src/ApiContext.tsx
@@ -1,0 +1,149 @@
+import React, { createContext, useState, useMemo, useEffect, use } from "react"
+import { ChatModels } from "../../core/src/chattypes"
+import {
+    Project,
+    ServerEnvResponse,
+    RunResultListResponse,
+} from "../../core/src/server/messages"
+import { fetchEnv, fetchScripts, fetchRuns, fetchModels } from "./api"
+import { useLocationHashValue } from "./useLocationHashValue"
+import { useUrlSearchParams } from "./useUrlSearchParam"
+import { FileWithPath } from "react-dropzone/."
+
+export type ImportedFile = FileWithPath & { selected?: boolean }
+
+export const ApiContext = createContext<{
+    project: Promise<Project | undefined>
+    env: Promise<ServerEnvResponse | undefined>
+
+    scriptid: string | undefined
+    setScriptid: (id: string) => void
+    files: string[]
+    setFiles: (files: string[]) => void
+    importedFiles: ImportedFile[]
+    setImportedFiles: (files: ImportedFile[]) => void
+    parameters: PromptParameters
+    setParameters: (parameters: PromptParameters) => void
+    options: ModelOptions
+    setOptions: (
+        f: (prev: ModelConnectionOptions) => ModelConnectionOptions
+    ) => void
+    refresh: () => void
+    runs: Promise<RunResultListResponse | undefined>
+    models: Promise<ChatModels | undefined>
+} | null>(null)
+
+export function ApiProvider({ children }: { children: React.ReactNode }) {
+    const [refreshId, setRefreshId] = useState(0)
+    const env = useMemo<Promise<ServerEnvResponse>>(fetchEnv, [refreshId])
+    const project = useMemo<Promise<Project>>(fetchScripts, [refreshId])
+    const runs = useMemo<Promise<RunResultListResponse>>(fetchRuns, [refreshId])
+    const models = useMemo<Promise<ChatModels>>(fetchModels, [refreshId])
+    const [scriptid, setScriptid] = useLocationHashValue("scriptid")
+
+    const refresh = () => setRefreshId((prev) => prev + 1)
+
+    const [state, setState] = useUrlSearchParams<
+        {
+            files: string[]
+        } & ModelConnectionOptions
+    >(
+        {
+            files: [],
+        },
+        {
+            scriptid: { type: "string" },
+            files: { type: "array", items: { type: "string" } },
+            cache: { type: "boolean" },
+            provider: { type: "string" },
+            model: { type: "string" },
+            smallModel: { type: "string" },
+            visionModel: { type: "string" },
+            temperature: { type: "number" },
+            logprobs: { type: "boolean" },
+            topLogprobs: { type: "integer" },
+        }
+    )
+    const [importedFiles, setImportedFiles] = useState<ImportedFile[]>([])
+    const { files, ...options } = state
+    const [parameters, setParameters] = useState<PromptParameters>({})
+    const setFiles = (files: string[]) =>
+        setState((prev) => ({
+            ...prev,
+            files: files.filter((s) => s !== "").slice(),
+        }))
+    const setOptions = (
+        f: (prev: ModelConnectionOptions) => ModelConnectionOptions
+    ) => {
+        setState((prev) => ({ ...prev, ...f(options) }))
+    }
+
+    return (
+        <ApiContext.Provider
+            value={{
+                project,
+                env,
+                scriptid,
+                setScriptid,
+                files,
+                setFiles,
+                importedFiles,
+                setImportedFiles,
+                parameters,
+                setParameters,
+                options,
+                setOptions,
+                refresh,
+                runs,
+                models,
+            }}
+        >
+            {children}
+        </ApiContext.Provider>
+    )
+}
+
+export function useApi() {
+    const api = use(ApiContext)
+    if (!api) throw new Error("missing content")
+    return api
+}
+
+export function useEnv() {
+    const { env: envPromise } = useApi()
+    const env = use(envPromise)
+    return env
+}
+
+export function useRunResults() {
+    const { runs: runsPromise } = useApi()
+    const runs = use(runsPromise)
+    return runs
+}
+
+export function useModels() {
+    const { models: modelsPromise } = useApi()
+    const models = use(modelsPromise)
+    return models
+}
+
+export function useProject() {
+    const api = useApi()
+    const project = use(api.project)
+    return project
+}
+
+export function useScripts() {
+    const project = useProject()
+    const scripts = (
+        project?.scripts?.filter((s) => !s.isSystem && !s.unlisted) || []
+    ).sort((l, r) => l.id.localeCompare(r.id))
+    return scripts
+}
+
+export function useScript() {
+    const scripts = useScripts()
+    const { scriptid } = useApi()
+
+    return scripts.find((s) => s.id === scriptid)
+}
